Added DB Connectivity and Houses
Created DB helper to support connection to a SQLite DB local to the device. Helper will create a SQLite DB if none exists and add the House table. Houses with their list of chores will now be stored in the DB.

Added Houses which have name, location, and list of chores intended for a group of users. Added activity to select from the list of houses. Added activity to add a new house.